decouple node-commands from cli-utils
