+++ toolkit/cli/node-commands/Cargo.toml
@@ -7,7 +7,6 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 [dependencies]
 authority-selection-inherents = { workspace = true }
 clap = { workspace = true }
-cli-utils = { workspace = true }
 env_logger = { workspace = true }
 frame-support = { workspace = true }
 log = { workspace = true }
@@ -32,8 +31,8 @@ serde_json = { workspace = true }
 ed25519-zebra = { workspace = true }
 secp256k1 = { workspace = true, features = ["std", "global-context"] }
 pallet-address-associations = { workspace = true, features = ["std"] }
+hex = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true }
-hex = { workspace = true }
 hex-literal = { workspace = true }
