+++ toolkit/cli/utils/src/key_params.rs
@@ -7,25 +7,6 @@ use std::io::ErrorKind;
 use std::str::FromStr;
 
 #[derive(Clone, Debug)]
-pub struct SidechainSigningKeyParam(pub SecretKey);
-
-impl SidechainSigningKeyParam {
-	pub fn to_pub_key(&self) -> PublicKey {
-		PublicKey::from_secret_key_global(&self.0)
-	}
-}
-
-impl FromStr for SidechainSigningKeyParam {
-	type Err = secp256k1::Error;
-
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		let trimmed = s.trim_start_matches("0x");
-		let pair = SecretKey::from_str(trimmed)?;
-		Ok(SidechainSigningKeyParam(pair))
-	}
-}
-
-#[derive(Clone, Debug)]
 pub struct SidechainPublicKeyParam(pub SidechainPublicKey);
 
 impl Display for SidechainPublicKeyParam {
@@ -96,28 +77,5 @@ impl FromStr for StakePoolSigningKeyParam {
 impl From<[u8; 32]> for StakePoolSigningKeyParam {
 	fn from(key: [u8; 32]) -> Self {
 		Self(ed25519_zebra::SigningKey::from(key))
-	}
-}
-
-impl StakePoolSigningKeyParam {
-	pub fn vkey(&self) -> StakePoolPublicKey {
-		StakePoolPublicKey(ed25519_zebra::VerificationKey::from(&self.0).into())
-	}
-}
-
-#[derive(Clone, Debug)]
-pub struct StakeSigningKeyParam(pub ed25519_zebra::SigningKey);
-
-impl FromStr for StakeSigningKeyParam {
-	type Err = Ed25519SigningKeyError;
-
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		Ok(Self(parse_zebra_signing_key(s)?))
-	}
-}
-
-impl StakeSigningKeyParam {
-	pub fn vkey(&self) -> StakePublicKey {
-		StakePublicKey(ed25519_zebra::VerificationKey::from(&self.0).into())
 	}
 }
